+++ examples/poyo/train.py
@@ -65,7 +65,6 @@ def main(cfg: DictConfig):
         num_latents_per_step=cfg.model.num_latents,
         modality_spec=modality_spec,
         sequence_length=cfg.sequence_length,
-        subtask_weights=cfg.subtask_weights,
     )
 
     # setup data module