+++ torch_brain/models/poyo.py
@@ -19,6 +19,7 @@ from torch_brain.registry import ModalitySpec
 from torch_brain.utils import (
     create_linspace_latent_tokens,
     create_start_end_unit_tokens,
+    resolve_weights_based_on_interval_membership,
 )
 
 
@@ -241,7 +242,6 @@ class POYOTokenizer:
         weight_registry (Dict): Registry of the weights.
         latent_step (float): Step size for generating latent tokens.
         num_latents_per_step (int): Number of latents per step.
-        subtask_weights (List[float]): Loss-weights for different subtasks.
     """
 
     def __init__(
@@ -253,13 +253,11 @@ class POYOTokenizer:
         modality_spec: ModalitySpec,
         sequence_length: float = 1.0,
         eval: bool = False,
-        subtask_weights: Optional[Iterable[float]] = None,
     ):
         self.unit_tokenizer = unit_tokenizer
         self.session_tokenizer = session_tokenizer
 
         self.modality_spec = modality_spec
-        self.subtask_weights = subtask_weights
 
         self.latent_step = latent_step
         self.num_latents_per_step = num_latents_per_step
@@ -308,33 +306,13 @@ class POYOTokenizer:
         if output_values.dtype == np.float64:
             output_values = output_values.astype(np.float32)
 
-        session_index = self.session_tokenizer(data.session)
-        session_index = np.repeat(session_index, len(output_timestamps))
+        output_session_index = self.session_tokenizer(data.session)
+        output_session_index = np.repeat(output_session_index, len(output_timestamps))
 
-        # Assign each output timestamp a movement_phase
-        movement_phase_assignment = dict()
-        for k in data.movement_phases.keys():
-            movement_phase_assignment[k] = interval_contains(
-                data.movement_phases.__dict__[k],
-                output_timestamps,
-            )
-
-        # Weights for the output predictions (used in the loss function)
-        if self.subtask_weights is None:
-            output_weights = np.ones(len(output_values), dtype=np.float32)
-        else:
-            output_weights = np.zeros_like(output_timestamps, dtype=np.float32)
-            for k in data.movement_phases.keys():
-                output_weights = output_weights + (
-                    movement_phase_assignment[k].astype(float) * self.subtask_weights[k]
-                )
-
-        # Mask for the output predictions
-        output_mask = np.ones(len(output_values), dtype=bool)
-        if self.eval:
-            # During eval, only evaluate on the subtask specified in the config
-            eval_movement_phase = data.config["eval_movement_phase"]
-            output_mask = output_mask & (movement_phase_assignment[eval_movement_phase])
+        # resolve weights
+        output_weights = resolve_weights_based_on_interval_membership(
+            output_timestamps, data, config=data.config.get("weights", None)
+        )
 
         batch = {
             # input sequence
@@ -346,9 +324,9 @@ class POYOTokenizer:
             "latent_index": latent_index,
             "latent_timestamps": latent_timestamps,
             # output sequence
-            "output_session_index": pad8(session_index),
+            "output_session_index": pad8(output_session_index),
             "output_timestamps": pad8(output_timestamps),
-            "output_mask": pad8(output_mask),
+            "output_mask": track_mask8(output_session_index),
             # ground truth targets
             "target_values": pad8(output_values),
             "target_weights": pad8(output_weights),
@@ -359,28 +337,3 @@ class POYOTokenizer:
             batch["absolute_start"] = data.absolute_start
 
         return batch
-
-
-def interval_contains(
-    interval: Interval, x: Union[float, np.ndarray]
-) -> Union[bool, np.ndarray]:
-    r"""If x is a single number, returns True if x is in the interval.
-    If x is a 1d numpy array, return a 1d bool numpy array.
-    """
-
-    if isinstance(x, float):
-        if len(interval) == 0:
-            return False
-
-        return np.logical_and(x >= interval.start, x < interval.end).any()
-    elif isinstance(x, np.ndarray):
-        if len(interval) == 0:
-            return np.zeros_like(x, dtype=bool)
-
-        x_expanded = x[:, None]
-        y = np.logical_and(
-            x_expanded >= interval.start[None, :], x_expanded < interval.end[None, :]
-        )
-        y = np.logical_or.reduce(y, axis=1)
-        assert y.shape == x.shape
-        return y